pyproject.toml
[tool.poetry]
name = "py_universal_loader"
version = "0.1.0"
description = "A unified, high-performance Python package for bulk-loading pandas DataFrames into diverse SQL databases and data warehouses.It provides a single, configuration-driven API that abstracts the most efficient native bulk-loading methods (e.g., `COPY`, `BULK INSERT`) for each backend."
authors = ["Gowtham A Rao <gowtham.rao@coreason.ai>"]
license = "Prosperity-3.0"
readme = "README.md"
packages = [{include = "py_universal_loader", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.12"
loguru = "^0.7.2"
pandas = "^2.3.3"
duckdb = "^1.4.1"
pyarrow = "^22.0.0"
psycopg2-binary = "^2.9.11"
boto3 = "^1.40.60"
google-cloud-bigquery = "^3.38.0"
snowflake-connector-python = {extras = ["pandas"], version = "^4.0.0"}
mysql-connector-python = "^9.5.0"
pyodbc = "^5.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
ruff = "^0.4.8"
pre-commit = "^3.7.1"
pytest-cov = "^5.0.0"
mkdocs = "^1.6.0"
mkdocs-material = "^9.5.26"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project]
name = "py_universal_loader"
version = "0.1.0"
description = "A unified, high-performance Python package for bulk-loading pandas DataFrames into diverse SQL databases and data warehouses.It provides a single, configuration-driven API that abstracts the most efficient native bulk-loading methods (e.g., `COPY`, `BULK INSERT`) for each backend."
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Gowtham A Rao", email = "gowtham.rao@coreason.ai" },
]
license = { file = "LICENSE" }
classifiers = [
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/CoReason-AI/py_universal_loader"
Repository = "https://github.com/CoReason-AI/py_universal_loader"
Documentation = "https://github.com/CoReason-AI/py_universal_loader"
